// src/app/Models/Candidato.model.ts
export interface Candidato {
    Id: string,
    EstatusCandidadoId: string;
    AniosTrabajando: number;
    PrestacionesEsperadas: string;
    ModalidadDeTrabajoId: number;
    ExpectativaSalarialNeta: number;
    CostoTotalAnual: number;
    NivelDeInglesId: number;
    ComentarioEntrevista: string;
    UltimoGradoDeEstudios: string;
    SaldoActual: number;
    CorreoPersonal: string;
    Direccion: string;
    FechaDelNacimiento: string;
    Nombre: string;
    Telefono: string;
  }
  